Django_AppBoletin/apps/webpublic/serializers.py
from rest_framework.serializers import ModelSerializer

from apps.webadmin.models import BoletinModelo

class BoletinSerializer(ModelSerializer):
	class Meta:
		model = BoletinModelo
		fields = ('titulo','resumen','tipo')

