Django_AppBoletin/requirements.txt
Django~=2.2.1
Djangorestframework~=3.9.4
Jinja2~=2.10.1